refactor(viewer): extract viewer_match.jsx from viewer.jsx
Pure file split, no behavior change. Step 4 of mp-pxxc.
